separando rotas - inicio da arquitetura de software
